refactor(web,api): pure api package not relying on env for trpc/query clients
