+++ apps/web/src/routes/_protected/posts/$postid/index.tsx
@@ -1,7 +1,8 @@
 import { ArrowLeftIcon, ReloadIcon } from '@radix-ui/react-icons';
 import { Button } from '@repo/ui/components/button';
 import { createFileRoute, Link } from '@tanstack/react-router';
-import { queryClient, trpc } from '@/router';
+import { queryClient } from '@/clients/queryClient';
+import { trpc } from '@/router';
 import { postsLinkOptions } from '@/validations/posts-link-options';
 
 export const Route = createFileRoute('/_protected/posts/$postid/')({
